move cloud log types and type checkers to separate file
