+++ packages/monitoring-system/data-processor/src/types/firestore-schema.ts
